requirements.txt
stable-baselines3==1.7.0
rl_zoo3==1.6.3
scikit-learn==0.24.2
imbalanced-learn==0.8.1
pillow==8.4.0
matplotlib==3.7.4
joblib==1.3.2
tqdm==4.62.3
statsmodels==0.14.0
pyglet==1.5.27
pandas==1.3.5
highway-env==1.5
sbx-rl==0.5.0
numpy==1.22.0
optax==0.1.7
seaborn==0.13.1
rliable==1.0.8
Box2D==2.3.10
glfw==2.6.5
imageio==2.33.1
mujoco==2.2.0